Refactor to allow tests to work: - Removed InProgressCommonVM, DesignVM inherits from InProgressVM directly - Added parameterless constructor for design/tests version. - Have a protected DesignTimeDownloadTasks SourceList that can be updated from the DesignVM - Bind Tasks to DesignTimeDownloadTasks in the design time parameterless constructor. - This is to avoid Tasks changing value in the DesignVM, which would break all bindings to it or require messy nested subscriptions.
